Improve URL joining behavior.
Doing a simple concatenation of the golink suffix with the base URL has
some issues. If the base URL doesn't contain any path, then the suffix
is appended directly onto the netloc portion which is likely not what
was intended. To fix this we first join the base URL's suffix with
the provided suffix and then join that to the scheme and netloc using
urlunsplit.

Moved Golink into the model module, added some validation and added
unit tests.