Move edit view to `/+edit`
This is somewhat nicer looking than the `?edit` query string and tends
to fit nicer into aiohttp's route/URL framework.

`+` is a URL safe character typically used as a replacement for space
(as an alternative to `%20` encoding). However it's unlikely to be used
at the start of a name, thus safe for us to reserve.