Keep track of Golink visits.
A visit is simply the number of times someone uses the Golink to be
redirected to a URL. Being redirected to +edit doesn't count.

This is also used as the ordering metric for showing results on the
+search page.

Renamed the database functions since everything is obviously focused on
Golinks, so no need to namespace.