Make database calls asynchronous.
There is no guarantee that database calls will happen immediately,
so to avoid blocking the main loop we should run all database calls in
a thread pool.

SQLite doesn't guarantee that it's thread safe, so we need to ensure
it's always run from the same thread. The easiest approach is a single
worker thread pool, though we could have also just created a single
worker thread.