Cleanup parsing of paths.
There is now only a single `GolinkView` for handling Golinks. This will
handle any path that doesn't start with a `/+`.

The name and suffix are now parsed from the path on first access to
their respective property attributes. These are then stored on the
request object for efficiency.

Invalid Golink names are now handled inside the view's handler rather
than as part of the route.

Move `/+static` handler first to ensure that it's always available.